% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_api_funs.r
\name{build_gov_query}
\alias{build_gov_query}
\title{Build Gov Query}
\usage{
build_gov_query(
  q = "",
  count = 10,
  start = 0,
  order = "-popularity",
  org_any = NULL,
  org_all = NULL,
  fields = NULL
)
}
\arguments{
\item{q}{query string}

\item{count}{(single integer) Maximum number of search results to return. If insufficient documents match, as many as possible are returned (subject to the supplied start offset). This may be set to 0 to return no results (which may be useful if only, say, facet values are wanted). Setting this to 0 will reduce processing time.}

\item{start}{(single integer) Position in search result list to start returning results (0-based) If the start offset is greater than the number of matching results, no results will be returned (but also no error will be returned).}

\item{order}{(single string) The sort order. A field name, with an optional preceding "-" to sort in descending order. If not specified, sort order is relevance. Only some fields can be sorted on - an HTTP 422 error will be returned if the requested field is not a valid sort field.}

\item{org_any}{list of organisation names to filter to. returns results if _any_ of the organisations match}

\item{org_all}{list of organisation names to filter to. returns results if _all_ of the organisations match}

\item{fields}{a vector of field names to return. See https://docs.publishing.service.gov.uk/apis/search/fields.html}
}
\value{

}
\description{
A function designed to make it simple to build query urls for the search API
}
\examples{
build_gov_query()
}
