% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/content_api_funs.r
\name{build_taxon_df}
\alias{build_taxon_df}
\title{Build Taxon Dataframe}
\usage{
build_taxon_df(
  include_world = TRUE,
  filter_world_help_services = TRUE,
  max_levels = Inf
)
}
\arguments{
\item{include_world}{include the world taxonomy tree? i.e. taxons below
world/all}

\item{filter_world_help_services}{Filter the world taxon to remove "help & services in}

\item{max_levels}{maximum number of levels to go down the taxon}
}
\value{

}
\description{
A wrapper function around `get_taxon_df` which works from the starting taxons
(gov.uk & gov.uk/world/all) and then recursively on child taxons to build a
complete taxonomy tree dataframe.
}
\examples{

build_taxon_df(max_levels = 2)
}
