% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_api_funs.r
\name{execute_gov_query}
\alias{execute_gov_query}
\title{Execute Gov Query}
\usage{
execute_gov_query(query_url)
}
\arguments{
\item{query_url}{A query url built by build_gov_query}
}
\value{
a dataframe of query results
}
\description{
A function which takes urls made using `build_gov_query`, executes the
resulting query, then formats the result as a dataframe
}
\details{
Note - this will result in some columns which have 0 length vectors in some of the rows, where the page does not have that attribute
These can be hard to work with - you can filter them out using something like:
dplyr::filter(purrr::map(<your-col-name>, is.list) == TRUE)
}
\examples{

execute_gov_query(build_gov_query())
}
