% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_col_funs.r
\name{filter_list_col}
\alias{filter_list_col}
\title{Filter List Column}
\usage{
filter_list_col(.data, column, items)
}
\arguments{
\item{column}{the list column}

\item{items}{a list of items to look for in the list column}

\item{df}{dataframe with a list column}
}
\value{

}
\description{
This filter works for a list column by creating an intermediary logical vector.
}
\examples{

related_pages <- 
  filter_list_col(
    .data = pages_df, 
    column = taxons, 
    items = taxon_group
  )

}
