% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visnetworks.r
\name{gen_visnetwork}
\alias{gen_visnetwork}
\title{Generate visnetwork}
\usage{
gen_visnetwork(
  network_df,
  threshold_field = n_pages_agg,
  threshold_value = 0,
  size_field = NULL,
  tooltip_field = "title",
  width = "1000",
  height = "600",
  hierarchical = FALSE,
  igraph_init = FALSE,
  seed = 123,
  smooth_edges = TRUE,
  ...
)
}
\arguments{
\item{network_df}{dataframe to turn into a network}

\item{threshold_field}{field to be used to prune the network}

\item{threshold_value}{value to be used in the `threshold_field` to prune the network}

\item{size_field}{field to use as the network size}

\item{tooltip_field}{field to use as hover text on the nodes}

\item{hierarchical}{should the nodes be laid out in a heirachical way}

\item{igraph_init}{use igraph package to initialise positions - faster but uglier}

\item{seed}{set the random seed for layout consistency}

\item{smooth_edges}{should edges be rendered as bezier curves - slower but results in nicer layout.}

\item{...}{additional arguments passed to the call to visNetwork}
}
\value{

}
\description{
Generate visnetwork
}
\examples{
gen_visnetwork(
   network_df = taxonomy_df_lower_threshold,
   hierarchical = TRUE 
)
}
