% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/content_api_funs.r
\name{get_taxon_df}
\alias{get_taxon_df}
\title{Get Taxon Dataframe}
\usage{
get_taxon_df(path = "https://www.gov.uk/api/content/")
}
\arguments{
\item{path}{content api path to query - i.e. www.gov.uk/api/content/<page>}
}
\value{
a dataframe of information about the taxon
}
\description{
Function to take a page and get the taxons it relates to
}
\examples{
get_taxon_df("https://www.gov.uk/api/content/")
}
