% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_funs.r
\name{list_up_tree}
\alias{list_up_tree}
\title{List Up Tree}
\usage{
list_up_tree(node, property)
}
\arguments{
\item{node}{data.tree node, for which a specified quantity is aggregated}

\item{property}{name of a property to be aggregated}
}
\value{
a data.tree with <property>_agg as a property
}
\description{
Recursive function to be applied to a tree node,
will aggregate a specified *list* from all child nodes, and grandchildren etc
}
\details{
Works as a side effect, output does not need to be assigned
}
\examples{

list_up_tree(node = taxonomy_tree, property = "name")
}
