% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_col_funs.r
\name{make_named_taxon_list}
\alias{make_named_taxon_list}
\title{Make Taxon Named List}
\usage{
make_named_taxon_list(taxon_df)
}
\arguments{
\item{taxon_df}{a taxon dataframe, including a name_agg column and a to
column. These are created by using `data.tree::ToDataFrameNetwork` on a
data.tree which has had the name (taxon) field aggregated using the
`list_up_tree` function. See `generate_trees.R` in data-raw folder for example.}
}
\value{

}
\description{
little wrapper around setNames to make a named list suitable for
`summary_by_taxon()`
}
\examples{
make_named_taxon_list(taxonomy_df_lower_threshold)
}
