% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_api_funs.r
\name{paginate_query}
\alias{paginate_query}
\title{Paginate Query}
\usage{
paginate_query(query_url, chunksize = 10, limit = 45)
}
\arguments{
\item{query_url}{A query url built by build_gov_query}

\item{chunksize}{number of results to get with each query}

\item{limit}{an upper limit on the total number of results to collect (rounded to chunksize)}
}
\value{

}
\description{
Paginate Query
}
\examples{
query_url <- build_gov_query()
paginate_query(query_url = query_url, chunksize = 3, limit = 10)
 
}
