% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_funs.r
\name{prepare_taxonomy_df}
\alias{prepare_taxonomy_df}
\title{Prepare Taxonomy Dataframe}
\usage{
prepare_taxonomy_df(taxon_df)
}
\arguments{
\item{taxon_df}{taxons dataframe as generated by `build_taxon_df()`}
}
\value{

}
\description{
Prepare Taxonomy Dataframe
}
\examples{
prepare_taxonomy_df(taxon_df)
}
