% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/content_api_funs.r
\name{safe_map}
\alias{safe_map}
\title{Safe Mapper}
\usage{
safe_map(.x, .f)
}
\arguments{
\item{.x}{an item to map over}

\item{.f}{a function to map which produces a dataframe / tibble}
}
\value{
a dataframe containing the results of mapping the function
}
\description{
a wrapper for map, that uses safely to bypass any errors and raises a warning
}
\examples{

safe_map(c("https://www.gov.uk/api/content/"), get_taxon_df)
}
