% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_col_funs.r
\name{summary_by_taxon_groups}
\alias{summary_by_taxon_groups}
\title{Summary by Taxon Groups}
\usage{
summary_by_taxon_groups(page_data, taxon_col, named_taxon_list, ...)
}
\arguments{
\item{page_data}{a dataframe of page data, which includes}

\item{taxon_col}{}

\item{named_taxon_list}{}

\item{...}{}
}
\value{

}
\description{
A function, which given some page data, tagged by taxon (where each page may relate to multiple taxons) can calculate summaries for groups of taxons
}
\examples{
taxon_group_page_counts <-
summary_by_taxon_groups(
 page_data = pages_df, 
 taxon_col = taxons, 
 named_taxon_list = taxon_groups,
 n = n()
 ) 

}
