% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_api_funs.r
\name{url_with_page}
\alias{url_with_page}
\title{URL With Page function to modify a url to set count & start parameters -
useful when mapping over pages of query}
\usage{
url_with_page(url, count, start)
}
\arguments{
\item{url}{url to modify - as prepared by build_gov_query}

\item{count}{(single integer) Maximum number of search results to return. If
insufficient documents match, as many as possible are returned (subject to
the supplied start offset). This may be set to 0 to return no results
(which may be useful if only, say, facet values are wanted). Setting this
to 0 will reduce processing time.}

\item{start}{(single integer) Position in search result list to start
returning results (0-based) If the start offset is greater than the number
of matching results, no results will be returned (but also no error will be
returned).}
}
\value{
url with count and start modified
}
\description{
URL With Page function to modify a url to set count & start parameters -
useful when mapping over pages of query
}
\examples{
# get second 10 pages
url_with_page(build_gov_query(), count = 10, start = 10)
}
